Enhance link management with user-specific authorization and minor UI improvements
- Add user-specific authorization for link update and delete operations
- Remove unused imports and improve error handling in link routes
- Replace `onPress` with `onClick` in link-related components
- Pass session to CategoryBody component for potential authorization
- Add TypeScript ignore for switch-cell component

+++ app/api/link/[id]/route.ts
@@ -3,7 +3,7 @@ import Link from '@/models/Link';
 import { connectDB } from '@/lib/db';
 import { auth } from '@/auth';
 
-export const GET = auth(async function GET(request: any, context: any) {
+export const GET = auth(async function GET(_request: any, context: any) {
   try {
     await connectDB();
     const link = await Link.findById(context.params.id);
@@ -19,13 +19,24 @@ export const PUT = auth(async function PUT(request: any, context: any) {
     if (!request.auth?.user) {
       return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
     }
+
     await connectDB();
     const data = await request.json();
     data.modifiedBy = request.auth.user.email;
-    const link = await Link.findByIdAndUpdate(context.params.id, data, {
+
+    const link = await Link.findById(context.params.id);
+    if (
+      request.auth.user.email !== link?.addedBy &&
+      request.auth.user.role !== 'admin'
+    ) {
+      return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+    }
+
+    const updatedLink = await Link.findByIdAndUpdate(context.params.id, data, {
       new: true
     });
-    return NextResponse.json(link);
+
+    return NextResponse.json(updatedLink);
   } catch (error) {
     console.error(error);
     return NextResponse.json({ message: 'An error occurred' }, { status: 500 });
@@ -37,10 +48,18 @@ export const DELETE = auth(async function DELETE(request: any, context: any) {
     if (!request.auth?.user) {
       return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
     }
-    console.log(context.params.id);
+
     await connectDB();
-    const link = await Link.findByIdAndDelete(context.params.id);
-    return NextResponse.json(link);
+
+    const link = await Link.findById(context.params.id);
+    if (
+      request.auth.user.email !== link?.addedBy &&
+      request.auth.user.role !== 'admin'
+    ) {
+      return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+    }
+    await Link.findByIdAndDelete(context.params.id);
+    return NextResponse.json({ message: 'Link deleted' });
   } catch (error) {
     console.error(error);
     return NextResponse.json({ message: 'An error occurred' }, { status: 500 });
